refactor: Cat, Item 도메인 설계 리팩토링
