+++ src/main/java/com/example/demo/domain/User.java
@@ -17,12 +17,6 @@ public class User {
     private String name;
 
     @Column(nullable = false)
-    private Integer point;
-
-    @Column
-    private String catName;
-
-    @Column(nullable = false)
     private String refreshToken;
 
     protected User() {
