+++ src/main/java/com/example/demo/domain/cat/ItemCategoryEnum.java
@@ -1,0 +1,5 @@
+package com.example.demo.domain.cat;
+
+public enum ItemCategoryEnum {
+    HAT
+}
