--- src/main/java/com/example/demo/domain/item/ItemCategoryEnum.java
@@ -1,5 +1,0 @@
-package com.example.demo.domain.item;
-
-public enum ItemCategoryEnum {
-    HAT, CLOTH, SHOES
-}
