--- src/main/java/com/example/demo/domain/item/UserItemId.java
@@ -1,34 +1,0 @@
-package com.example.demo.domain.item;
-
-import jakarta.persistence.Embeddable;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-@Embeddable
-public class UserItemId implements Serializable {
-    private Long userId;
-    private Long itemId;
-
-    protected UserItemId() {
-    }
-
-    public UserItemId(Long userId, Long itemId) {
-        this.userId = userId;
-        this.itemId = itemId;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        UserItemId that = (UserItemId) o;
-        return Objects.equals(userId, that.userId) &&
-                Objects.equals(itemId, that.itemId);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(userId, itemId);
-    }
-}
